requirements.txt
# ── Core AI / OpenAI ──────────────────────────────────────────────────────────
openai>=1.30.0
python-dotenv>=1.0.0

# ── LangChain (RAG pipeline) ──────────────────────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ── Vector search ─────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Lexical search ────────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ── Re-ranking ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ── FastAPI backend ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── MongoDB async driver ──────────────────────────────────────────────────────
motor>=3.4.0
pymongo>=4.7.0

# ── Web app (Streamlit frontend) ──────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0

# ── Visualisation (notebook only) ─────────────────────────────────────────────
matplotlib>=3.8.0
